+++ core/bin/merkle_tree_consistency_checker/src/main.rs
@@ -28,7 +28,7 @@ impl Cli {
         tracing::info!("Verifying consistency of Merkle tree at {db_path}");
         let start = Instant::now();
         let db = RocksDB::new(Path::new(db_path));
-        let tree = ZkSyncTree::new_lightweight(db);
+        let tree = ZkSyncTree::new_lightweight(db.into());
 
         let l1_batch_number = if let Some(number) = self.l1_batch {
             L1BatchNumber(number)