+++ core/lib/dal/src/prover_dal.rs
@@ -7,13 +7,9 @@ use std::{
 
 use sqlx::Error;
 use zksync_types::{
-    aggregated_operations::L1BatchProofForL1,
     proofs::{
         AggregationRound, JobCountStatistics, JobExtendedStatistics, ProverJobInfo,
         ProverJobMetadata,
-    },
-    zkevm_test_harness::{
-        abstract_zksync_circuit::concrete_circuits::ZkSyncProof, bellman::bn256::Bn256,
     },
     L1BatchNumber, ProtocolVersionId,
 };
@@ -37,22 +33,34 @@ impl ProverDal<'_, '_> {
     ) -> Option<ProverJobMetadata> {
         let protocol_versions: Vec<i32> = protocol_versions.iter().map(|&id| id as i32).collect();
         let result: Option<ProverJobMetadata> = sqlx::query!(
-            "
-                UPDATE prover_jobs
-                SET status = 'in_progress', attempts = attempts + 1,
-                    updated_at = now(), processing_started_at = now()
-                WHERE id = (
-                        SELECT id
-                        FROM prover_jobs
-                        WHERE status = 'queued'
-                        AND protocol_version = ANY($1)
-                        ORDER BY aggregation_round DESC, l1_batch_number ASC, id ASC
-                        LIMIT 1
-                        FOR UPDATE
+            r#"
+            UPDATE prover_jobs
+            SET
+                status = 'in_progress',
+                attempts = attempts + 1,
+                updated_at = NOW(),
+                processing_started_at = NOW()
+            WHERE
+                id = (
+                    SELECT
+                        id
+                    FROM
+                        prover_jobs
+                    WHERE
+                        status = 'queued'
+                        AND protocol_version = ANY ($1)
+                    ORDER BY
+                        aggregation_round DESC,
+                        l1_batch_number ASC,
+                        id ASC
+                    LIMIT
+                        1
+                    FOR UPDATE
                         SKIP LOCKED
                 )
-                RETURNING prover_jobs.*
-                ",
+            RETURNING
+                prover_jobs.*
+            "#,
             &protocol_versions[..]
         )
         .fetch_optional(self.storage.conn())
@@ -71,22 +79,29 @@ impl ProverDal<'_, '_> {
     pub async fn get_proven_l1_batches(&mut self) -> Vec<(L1BatchNumber, AggregationRound)> {
         {
             sqlx::query!(
-                r#"SELECT MAX(l1_batch_number) as "l1_batch_number!", aggregation_round FROM prover_jobs 
-                 WHERE status='successful'
-                 GROUP BY aggregation_round 
+                r#"
+                SELECT
+                    MAX(l1_batch_number) AS "l1_batch_number!",
+                    aggregation_round
+                FROM
+                    prover_jobs
+                WHERE
+                    status = 'successful'
+                GROUP BY
+                    aggregation_round
                 "#
             )
-                .fetch_all(self.storage.conn())
-                .await
-                .unwrap()
-                .into_iter()
-                .map(|record| {
-                    (
-                        L1BatchNumber(record.l1_batch_number as u32),
-                        record.aggregation_round.try_into().unwrap(),
-                    )
-                })
-                .collect()
+            .fetch_all(self.storage.conn())
+            .await
+            .unwrap()
+            .into_iter()
+            .map(|record| {
+                (
+                    L1BatchNumber(record.l1_batch_number as u32),
+                    record.aggregation_round.try_into().unwrap(),
+                )
+            })
+            .collect()
         }
     }
 
@@ -99,23 +114,35 @@ impl ProverDal<'_, '_> {
             let protocol_versions: Vec<i32> =
                 protocol_versions.iter().map(|&id| id as i32).collect();
             let result: Option<ProverJobMetadata> = sqlx::query!(
-                "
+                r#"
                 UPDATE prover_jobs
-                SET status = 'in_progress', attempts = attempts + 1,
-                    updated_at = now(), processing_started_at = now()
-                WHERE id = (
-                        SELECT id
-                        FROM prover_jobs
-                        WHERE circuit_type = ANY($1)
-                        AND status = 'queued'
-                        AND protocol_version = ANY($2)
-                        ORDER BY aggregation_round DESC, l1_batch_number ASC, id ASC
-                        LIMIT 1
+                SET
+                    status = 'in_progress',
+                    attempts = attempts + 1,
+                    updated_at = NOW(),
+                    processing_started_at = NOW()
+                WHERE
+                    id = (
+                        SELECT
+                            id
+                        FROM
+                            prover_jobs
+                        WHERE
+                            circuit_type = ANY ($1)
+                            AND status = 'queued'
+                            AND protocol_version = ANY ($2)
+                        ORDER BY
+                            aggregation_round DESC,
+                            l1_batch_number ASC,
+                            id ASC
+                        LIMIT
+                            1
                         FOR UPDATE
-                        SKIP LOCKED
+                            SKIP LOCKED
                     )
-                RETURNING prover_jobs.*
-                ",
+                RETURNING
+                    prover_jobs.*
+                "#,
                 &circuit_types[..],
                 &protocol_versions[..]
             )
@@ -146,11 +173,24 @@ impl ProverDal<'_, '_> {
             let it = circuit_types_and_urls.into_iter().enumerate();
             for (sequence_number, (circuit, circuit_input_blob_url)) in it {
                 sqlx::query!(
-                    "
-                    INSERT INTO prover_jobs (l1_batch_number, circuit_type, sequence_number, prover_input, aggregation_round, circuit_input_blob_url, protocol_version, status, created_at, updated_at)
-                    VALUES ($1, $2, $3, $4, $5, $6, $7, 'queued', now(), now())
-                    ON CONFLICT(l1_batch_number, aggregation_round, sequence_number) DO NOTHING
-                    ",
+                    r#"
+                    INSERT INTO
+                        prover_jobs (
+                            l1_batch_number,
+                            circuit_type,
+                            sequence_number,
+                            prover_input,
+                            aggregation_round,
+                            circuit_input_blob_url,
+                            protocol_version,
+                            status,
+                            created_at,
+                            updated_at
+                        )
+                    VALUES
+                        ($1, $2, $3, $4, $5, $6, $7, 'queued', NOW(), NOW())
+                    ON CONFLICT (l1_batch_number, aggregation_round, sequence_number) DO NOTHING
+                    "#,
                     l1_batch_number.0 as i64,
                     circuit,
                     sequence_number as i64,
@@ -180,11 +220,17 @@ impl ProverDal<'_, '_> {
     ) -> Result<(), Error> {
         {
             sqlx::query!(
-                "
+                r#"
                 UPDATE prover_jobs
-                SET status = 'successful', updated_at = now(), time_taken = $1, result = $2, proccesed_by = $3
-                WHERE id = $4
-                ",
+                SET
+                    status = 'successful',
+                    updated_at = NOW(),
+                    time_taken = $1,
+                    result = $2,
+                    proccesed_by = $3
+                WHERE
+                    id = $4
+                "#,
                 duration_to_naive_time(time_taken),
                 &proof,
                 proccesed_by,
@@ -210,12 +256,18 @@ impl ProverDal<'_, '_> {
             let mut transaction = self.storage.start_transaction().await.unwrap();
 
             let row = sqlx::query!(
-                "
+                r#"
                 UPDATE prover_jobs
-                SET status = 'failed', error = $1, updated_at = now()
-                WHERE id = $2
-                RETURNING l1_batch_number, attempts
-                ",
+                SET
+                    status = 'failed',
+                    error = $1,
+                    updated_at = NOW()
+                WHERE
+                    id = $2
+                RETURNING
+                    l1_batch_number,
+                    attempts
+                "#,
                 error,
                 id as i64,
             )
@@ -243,65 +295,45 @@ impl ProverDal<'_, '_> {
         let processing_timeout = pg_interval_from_duration(processing_timeout);
         {
             sqlx::query!(
-                "
+                r#"
                 UPDATE prover_jobs
-                SET status = 'queued', updated_at = now(), processing_started_at = now()
-                WHERE (status = 'in_progress' AND  processing_started_at <= now() - $1::interval AND attempts < $2)
-                OR (status = 'in_gpu_proof' AND  processing_started_at <= now() - $1::interval AND attempts < $2)
-                OR (status = 'failed' AND attempts < $2)
-                RETURNING id, status, attempts
-                ",
+                SET
+                    status = 'queued',
+                    updated_at = NOW(),
+                    processing_started_at = NOW()
+                WHERE
+                    (
+                        status = 'in_progress'
+                        AND processing_started_at <= NOW() - $1::INTERVAL
+                        AND attempts < $2
+                    )
+                    OR (
+                        status = 'in_gpu_proof'
+                        AND processing_started_at <= NOW() - $1::INTERVAL
+                        AND attempts < $2
+                    )
+                    OR (
+                        status = 'failed'
+                        AND attempts < $2
+                    )
+                RETURNING
+                    id,
+                    status,
+                    attempts
+                "#,
                 &processing_timeout,
                 max_attempts as i32,
             )
-                .fetch_all(self.storage.conn())
-                .await
-                .unwrap()
-                .into_iter()
-                .map(|row| StuckProverJobs{id: row.id as u64, status: row.status, attempts: row.attempts as u64})
-                .collect()
-        }
-    }
-
-    // For each block in the provided range it returns a tuple:
-    // (aggregation_coords; scheduler_proof)
-    pub async fn get_final_proofs_for_blocks(
-        &mut self,
-        from_block: L1BatchNumber,
-        to_block: L1BatchNumber,
-    ) -> Vec<L1BatchProofForL1> {
-        {
-            sqlx::query!(
-                "SELECT prover_jobs.result as proof, scheduler_witness_jobs.aggregation_result_coords
-                FROM prover_jobs
-                INNER JOIN scheduler_witness_jobs
-                ON prover_jobs.l1_batch_number = scheduler_witness_jobs.l1_batch_number
-                WHERE prover_jobs.l1_batch_number >= $1 AND prover_jobs.l1_batch_number <= $2
-                AND prover_jobs.aggregation_round = 3
-                AND prover_jobs.status = 'successful'
-                ",
-                from_block.0 as i32,
-                to_block.0 as i32
-            )
-                .fetch_all(self.storage.conn())
-                .await
-                .unwrap()
-                .into_iter()
-                .map(|row| {
-                    let deserialized_proof = bincode::deserialize::<ZkSyncProof<Bn256>>(
-                        &row.proof
-                            .expect("prove_job with `successful` status has no result"),
-                    ).expect("cannot deserialize proof");
-                    let deserialized_aggregation_result_coords = bincode::deserialize::<[[u8; 32]; 4]>(
-                        &row.aggregation_result_coords
-                            .expect("scheduler_witness_job with `successful` status has no aggregation_result_coords"),
-                    ).expect("cannot deserialize proof");
-                    L1BatchProofForL1 {
-                        aggregation_result_coords: deserialized_aggregation_result_coords,
-                        scheduler_proof: ZkSyncProof::into_proof(deserialized_proof),
-                    }
-                })
-                .collect()
+            .fetch_all(self.storage.conn())
+            .await
+            .unwrap()
+            .into_iter()
+            .map(|row| StuckProverJobs {
+                id: row.id as u64,
+                status: row.status,
+                attempts: row.attempts as u64,
+            })
+            .collect()
         }
     }
 
@@ -311,10 +343,18 @@ impl ProverDal<'_, '_> {
         {
             sqlx::query!(
                 r#"
-                SELECT COUNT(*) as "count!", circuit_type as "circuit_type!", status as "status!"
-                FROM prover_jobs
-                WHERE status <> 'skipped' and status <> 'successful' 
-                GROUP BY circuit_type, status
+                SELECT
+                    COUNT(*) AS "count!",
+                    circuit_type AS "circuit_type!",
+                    status AS "status!"
+                FROM
+                    prover_jobs
+                WHERE
+                    status <> 'skipped'
+                    AND status <> 'successful'
+                GROUP BY
+                    circuit_type,
+                    status
                 "#
             )
             .fetch_all(self.storage.conn())
@@ -345,9 +385,13 @@ impl ProverDal<'_, '_> {
         {
             let mut results: HashMap<String, usize> = sqlx::query!(
                 r#"
-                SELECT COUNT(*) as "count!", status as "status!"
-                FROM prover_jobs
-                GROUP BY status
+                SELECT
+                    COUNT(*) AS "count!",
+                    status AS "status!"
+                FROM
+                    prover_jobs
+                GROUP BY
+                    status
                 "#
             )
             .fetch_all(self.storage.conn())
@@ -369,13 +413,22 @@ impl ProverDal<'_, '_> {
         {
             sqlx::query!(
                 r#"
-                SELECT MIN(l1_batch_number) as "l1_batch_number?" FROM (
-                    SELECT MIN(l1_batch_number) as "l1_batch_number"
-                    FROM prover_jobs
-                    WHERE status = 'successful' OR aggregation_round < 3
-                    GROUP BY l1_batch_number
-                    HAVING MAX(aggregation_round) < 3
-                ) as inn
+                SELECT
+                    MIN(l1_batch_number) AS "l1_batch_number?"
+                FROM
+                    (
+                        SELECT
+                            MIN(l1_batch_number) AS "l1_batch_number"
+                        FROM
+                            prover_jobs
+                        WHERE
+                            status = 'successful'
+                            OR aggregation_round < 3
+                        GROUP BY
+                            l1_batch_number
+                        HAVING
+                            MAX(aggregation_round) < 3
+                    ) AS inn
                 "#
             )
             .fetch_one(self.storage.conn())
@@ -392,12 +445,21 @@ impl ProverDal<'_, '_> {
         {
             sqlx::query!(
                 r#"
-                    SELECT MIN(l1_batch_number) as "l1_batch_number!", circuit_type
-                    FROM prover_jobs
-                    WHERE aggregation_round = 0 AND (status = 'queued' OR status = 'in_progress'
-                    OR status = 'in_gpu_proof'
-                    OR status = 'failed')
-                    GROUP BY circuit_type
+                SELECT
+                    MIN(l1_batch_number) AS "l1_batch_number!",
+                    circuit_type
+                FROM
+                    prover_jobs
+                WHERE
+                    aggregation_round = 0
+                    AND (
+                        status = 'queued'
+                        OR status = 'in_progress'
+                        OR status = 'in_gpu_proof'
+                        OR status = 'failed'
+                    )
+                GROUP BY
+                    circuit_type
                 "#
             )
             .fetch_all(self.storage.conn())
@@ -414,19 +476,36 @@ impl ProverDal<'_, '_> {
             let limits = sqlx::query!(
                 r#"
                 SELECT
-                    (SELECT l1_batch_number
-                    FROM prover_jobs
-                    WHERE status NOT IN ('successful', 'skipped')
-                    ORDER BY l1_batch_number
-                    LIMIT 1) as "successful_limit!",
-                    
-                    (SELECT l1_batch_number
-                    FROM prover_jobs
-                    WHERE status <> 'queued'
-                    ORDER BY l1_batch_number DESC
-                    LIMIT 1) as "queued_limit!",
-
-                    (SELECT MAX(l1_batch_number) as "max!" FROM prover_jobs) as "max_block!"
+                    (
+                        SELECT
+                            l1_batch_number
+                        FROM
+                            prover_jobs
+                        WHERE
+                            status NOT IN ('successful', 'skipped')
+                        ORDER BY
+                            l1_batch_number
+                        LIMIT
+                            1
+                    ) AS "successful_limit!",
+                    (
+                        SELECT
+                            l1_batch_number
+                        FROM
+                            prover_jobs
+                        WHERE
+                            status <> 'queued'
+                        ORDER BY
+                            l1_batch_number DESC
+                        LIMIT
+                            1
+                    ) AS "queued_limit!",
+                    (
+                        SELECT
+                            MAX(l1_batch_number) AS "max!"
+                        FROM
+                            prover_jobs
+                    ) AS "max_block!"
                 "#
             )
             .fetch_one(self.storage.conn())
@@ -541,9 +620,19 @@ impl ProverDal<'_, '_> {
         job_id: u32,
     ) -> Result<Option<ProverJobMetadata>, Error> {
         {
-            let row = sqlx::query!("SELECT * from prover_jobs where id=$1", job_id as i64)
-                .fetch_optional(self.storage.conn())
-                .await?;
+            let row = sqlx::query!(
+                r#"
+                SELECT
+                    *
+                FROM
+                    prover_jobs
+                WHERE
+                    id = $1
+                "#,
+                job_id as i64
+            )
+            .fetch_optional(self.storage.conn())
+            .await?;
 
             Ok(row.map(|row| ProverJobMetadata {
                 id: row.id as u32,
@@ -562,11 +651,17 @@ impl ProverDal<'_, '_> {
         {
             let job_ids = sqlx::query!(
                 r#"
-                    SELECT id, circuit_input_blob_url FROM prover_jobs
-                    WHERE status='successful'
-                    AND circuit_input_blob_url is NOT NULL
+                SELECT
+                    id,
+                    circuit_input_blob_url
+                FROM
+                    prover_jobs
+                WHERE
+                    status = 'successful'
+                    AND circuit_input_blob_url IS NOT NULL
                     AND updated_at < NOW() - INTERVAL '30 days'
-                    LIMIT $1;
+                LIMIT
+                    $1;
                 "#,
                 limit as i32
             )
@@ -585,8 +680,11 @@ impl ProverDal<'_, '_> {
             sqlx::query!(
                 r#"
                 UPDATE prover_jobs
-                SET status = $1, updated_at = now()
-                WHERE id = $2
+                SET
+                    status = $1,
+                    updated_at = NOW()
+                WHERE
+                    id = $2
                 "#,
                 status,
                 id as i64,